create handler and token package
